server/requirements.txt
django
mongoengine
djangorestframework
